requirements.txt
matplotlib==3.2.2
opencv-python==4.2.0.34
pandas==1.0.4
Pillow==7.1.2
pydicom==2.0.0
